Removed share_controllers and moved code into share endpoint for groups and projects controllers. Removed share_controller tests and moved into groups_controller and projects_controller tests. Added translations. Updated routes
