+++ config/routes/group.rb
@@ -6,6 +6,7 @@ constraints(::Constraints::GroupUrlConstrainer.new) do
         constraints: { id: Irida::PathRegex.full_namespace_route_regex }) do
     scope(path: '-') do
       get :edit, as: :edit_group
+      post :share, as: :group_share
     end
 
     get '/', action: :show, as: :group_canonical
@@ -17,7 +18,6 @@ constraints(::Constraints::GroupUrlConstrainer.new) do
         constraints: { group_id: Irida::PathRegex.full_namespace_route_regex }) do
     resources :members, only: %i[create destroy index new update]
     resources :samples, only: %i[index]
-    post '/share', to: 'share#create'
   end
 
   scope(path: '*id',
