+++ config/routes/project.rb
@@ -10,7 +10,6 @@ constraints(::Constraints::ProjectUrlConstrainer.new) do
     patch '/', action: :update
     put '/', action: :update
     delete '/', action: :destroy
-    post '/share', to: 'share#create'
 
     # Begin on /-/ scope.
     # Use this for all project routes.
@@ -26,6 +25,7 @@ constraints(::Constraints::ProjectUrlConstrainer.new) do
           end
         end
       end
+      post :share
     end
   end
 end
